requirements.txt
groq
hume
streamlit
python-dotenv
langchain
langchain-core
langchain-groq
crewai
crewai[tools]
browserbase